requirements.txt
python-dotenv
requests
python-telegram-bot==20.3
